Add challenges, clean up unnecessary code
I added a Challenge domain object to represent a challenge made between
two players. The ChallengeService allows for business operations to be
performed on challenges, and the ChallengeDeprecationService is scheduled
to deprecated expiring challenges.
Also, I added a bunch of unit tests and cleaned up unused example code.
This is definitely way too much for a single commit, but it represents the initial
work I've done for this project and I got a bit carried away.